network final tcp fix + libc basics

+++ include/net/tcp.h
@@ -9,26 +9,26 @@ namespace guyos
 {
     namespace net
     {
-
-
+        
+        
         enum TransmissionControlProtocolSocketState
         {
             CLOSED,
             LISTEN,
             SYN_SENT,
             SYN_RECEIVED,
-
+            
             ESTABLISHED,
-
+            
             FIN_WAIT1,
             FIN_WAIT2,
             CLOSING,
             TIME_WAIT,
-
+            
             CLOSE_WAIT
             //LAST_ACK
         };
-
+        
         enum TransmissionControlProtocolFlag
         {
             FIN = 1,
@@ -41,8 +41,8 @@ namespace guyos
             CWR = 128,
             NS = 256
         };
-
-
+        
+        
         struct TransmissionControlProtocolHeader
         {
             common::uint16_t srcPort;
@@ -52,15 +52,16 @@ namespace guyos
             
             common::uint8_t reserved : 4;
             common::uint8_t headerSize32 : 4;
-            common::uint16_t flags;
-
+            common::uint8_t flags;
+            
             common::uint16_t windowSize;
             common::uint16_t checksum;
             common::uint16_t urgentPtr;
-
+            
             common::uint32_t options;
         } __attribute__((packed));
-
+       
+      
         struct TransmissionControlProtocolPseudoHeader
         {
             common::uint32_t srcIP;
@@ -68,13 +69,13 @@ namespace guyos
             common::uint16_t protocol;
             common::uint16_t totalLength;
         } __attribute__((packed));
-
-
+      
+      
         class TransmissionControlProtocolSocket;
         class TransmissionControlProtocolProvider;
-
-
-
+        
+        
+        
         class TransmissionControlProtocolHandler
         {
         public:
@@ -82,9 +83,9 @@ namespace guyos
             ~TransmissionControlProtocolHandler();
             virtual bool HandleTransmissionControlProtocolMessage(TransmissionControlProtocolSocket* socket, common::uint8_t* data, common::uint16_t size);
         };
-
-
-
+      
+        
+      
         class TransmissionControlProtocolSocket
         {
         friend class TransmissionControlProtocolProvider;
@@ -95,9 +96,10 @@ namespace guyos
             common::uint32_t localIP;
             common::uint32_t sequenceNumber;
             common::uint32_t acknowledgementNumber;
+
             TransmissionControlProtocolProvider* backend;
             TransmissionControlProtocolHandler* handler;
-
+            
             TransmissionControlProtocolSocketState state;
         public:
             TransmissionControlProtocolSocket(TransmissionControlProtocolProvider* backend);
@@ -106,19 +108,19 @@ namespace guyos
             virtual void Send(common::uint8_t* data, common::uint16_t size);
             virtual void Disconnect();
         };
-
-
+      
+      
         class TransmissionControlProtocolProvider : InternetProtocolHandler
         {
         protected:
             TransmissionControlProtocolSocket* sockets[65535];
             common::uint16_t numSockets;
             common::uint16_t freePort;
-
+            
         public:
             TransmissionControlProtocolProvider(InternetProtocolProvider* backend);
             ~TransmissionControlProtocolProvider();
-
+            
             virtual bool OnInternetProtocolReceived(common::uint32_t srcIP_BE, common::uint32_t dstIP_BE,
                                                     common::uint8_t* internetprotocolPayload, common::uint32_t size);
 
@@ -130,9 +132,6 @@ namespace guyos
             virtual TransmissionControlProtocolSocket* Listen(common::uint16_t port);
             virtual void Bind(TransmissionControlProtocolSocket* socket, TransmissionControlProtocolHandler* handler);
         };
-
-
-
 
     }
 }
